refactor: session 및 request 패키지 변경
