+++ tomcat/src/main/java/org/apache/coyote/handler/mapping/RegisterMapping.java
@@ -2,7 +2,7 @@ package org.apache.coyote.handler.mapping;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http.HttpRequest;
+import org.apache.coyote.http.request.HttpRequest;
 
 import java.io.IOException;
 import java.util.Map;
