+++ tomcat/src/main/java/org/apache/coyote/http/LoginManager.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.http;
 
+import org.apache.coyote.http.session.Session;
+
 public interface LoginManager {
 
     void add(final Session session);
