+++ tomcat/src/main/java/org/apache/coyote/http/request/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http;
+package org.apache.coyote.http.request;
 
 import java.util.Map;
